MAISTRA-2516: Consider values from global.yaml when rendering charts
Currently there's a bug in the code that renders charts that it's not
correctly merging the fields from global.yaml in the resulting chart.
The field status.appliedValues is also affected by this bug.

We fix the bug by setting back `spec.Istio` with the result of
the merge of its previous content and global.yaml.